Store GSR as Float in Room entity

+++ app/src/main/java/pl/edu/pjwstk/engineeringthesis/data/db/entity/GsrSampleEntity.kt
@@ -13,7 +13,7 @@ data class GsrSampleEntity(
 
     val epoch : Long,
 
-    val gsr : Int
+    val gsr : Float
 ){
     fun toDomain() : GsrSample{
         return GsrSample(
